python-api-testing/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
pymongo==4.6.1
pydantic==2.6.1
python-multipart==0.0.9
aiohttp==3.9.3
jinja2==3.1.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.25.1
python-dotenv==1.0.0
structlog==23.2.0 